@@ -1,13 +1,9 @@
 {
-  "name": "Project PWA",
+  "name": "projectpwa",
   "version": "0.1.0",
   "private": true,
-  "engines": {
-    "node": "14.17.3"
-  },
   "dependencies": {
-    "cypress": "10.10.0",
-    "typescript": "^4.6.3"
+    "cypress": "10.10.0"
   },
   "scripts": {
     "cypress:test-headless": "cypress run --browser chrome --headless",
@@ -16,6 +12,7 @@
   "devDependencies": {
     "cypress-image-diff-js": "^1.22.0",
     "mocha": "^10.0.0",
-    "prettier": "^2.5.1"
+    "prettier": "^2.5.1",
+    "typescript": "^4.8.4"
   }
 }
