@@ -1,5 +1,5 @@
-import { mainPage } from '../locators';
-import utils from '../src/utils';
+import { selectors } from '../locators/Selectors';
+import { utils } from '../src/Utils';
 
 describe('home page tests', () => {
 	let accountToBeDeletedUid
@@ -11,9 +11,9 @@ describe('home page tests', () => {
 			accountToBeDeletedUid = data.body.uid
 			utils.createAccountAndSignIn(data.body)
 		})
-		cy.get(mainPage.profileLogo).click();
+		cy.get(selectors.mainPage.profileLogo).click();
 		cy.wait('@workspace');
-		cy.compareSnapshot('profile-page');
+		cy.compareSnapshot('profile-page', 0);
 	});
 
 	it('Verify clicking on "Add my first debt" button should take you to debt page, ID: 18', () => {
@@ -32,7 +32,7 @@ describe('home page tests', () => {
 			utils.createAccountAndSignIn(data.body);
 		});
 		cy.wait('@workspace')
-		cy.get(mainPage.primaryButton).click();
+		cy.get(selectors.mainPage.primaryButton).click();
 		cy.url().should('include', 'settings/workspaces');
 	});
 
@@ -42,7 +42,7 @@ describe('home page tests', () => {
 			utils.createAccountAndSignIn(data.body);
 		});
 		utils.fillDebtInfo('adewdbc', '2000', '300', '300');
-		cy.get(mainPage.navbarHomeButton).click();
+		cy.get(selectors.mainPage.navbarHomeButton).click();
 		cy.wait('@planner');
 		// wait for home page to load
 		cy.wait(5000);
