@@ -1,5 +1,5 @@
-export default {
-  loginPage: {
+class Selectors { 
+  loginPage = {
     backButton: 'a[class*=mat-button-base]',
     codeInput: '#code0',
     forgotPasswordLink: '[href="#/public/forgot-password"]',
@@ -9,22 +9,21 @@ export default {
     signinLink: '[href="#/public/sign-in"]',
     signupLink: '[href="#/public/sign-up"]',
     usernameInput: '[id*="input_username"]',
-  },
-  mainPage: {
+  };
+  mainPage = {
     addMyFirstDebtButton: '[class*="mat-f"]',
     primaryButton:'[class*="workspace-2"]',
     profileLogo: '[mattooltip="Profile"]',
     navbarHomeButton: '[data-mat-icon-name="logo_bird"]',
+  }
 
-  },
-
-  shared:{
+  shared = {
     body: 'body',
     submitButton: '[type="submit"]',
     validationMessage: 'formly-validation-message',
-  },
+  }
 
-  debtPage:{
+  debtPage = {
     addDebtButton: '[class^="mat-focus"]',
     annualPercentageRateInput :'[id*="input_no_promo"]',
     currentBalanceInput:'[id*="input_current_balance"]',
@@ -37,5 +36,6 @@ export default {
     previousMonthButton:'[aria-label="Previous month"]',
     calenderDaysButtons: '[class*="cell-content"]',
     oldDataMsg: 'simple-snack-bar > span',
-  },
-};
+  }
+}
+export const selectors = new Selectors();
