@@ -8,10 +8,10 @@ beforeEach(() => {
   cy.intercept('api/v2/user/workspace/*').as('workspace')
   cy.intercept('api/v2/user/data').as('user-data')
   cy.intercept('https://www.debtpayoffplanner.com//wp-json/wp/v2/*').as('planner');
-  
+
+  // const date = new Date('October 22, 2022').getTime();
+  // cy.clock(date)
 });
-
-
 
 after(() => {
   cy.task('generateReport')
